improve course page architecture with custom hooks and components
