+++ src/components/content/courses/details/DesktopCourseDetails.js
