+++ src/components/content/courses/details/DraftCourseDetails.js
