+++ src/components/content/courses/details/MobileCourseDetails.js
