+++ src/components/content/courses/layout/CourseHeader.js
