+++ src/components/content/courses/layout/CourseSidebar.js
