+++ src/components/content/courses/lessons/CombinedLesson.js
