+++ src/components/content/courses/lessons/DocumentLesson.js
