+++ src/pages/course/[slug]/draft/index.js
@@ -2,8 +2,8 @@ import React, { useEffect, useState, useCallback } from 'react';
 import { useRouter } from 'next/router';
 import axios from 'axios';
 import { parseEvent, findKind0Fields } from '@/utils/nostr';
-import DraftCourseDetails from '@/components/content/courses/DraftCourseDetails';
-import DraftCourseLesson from '@/components/content/courses/DraftCourseLesson';
+import DraftCourseDetails from '@/components/content/courses/details/DraftCourseDetails';
+import DraftCourseLesson from '@/components/content/courses/details/DraftCourseLesson';
 import { useNDKContext } from '@/context/NDKContext';
 import { useSession } from 'next-auth/react';
 import { useIsAdmin } from '@/hooks/useIsAdmin';
